increase maps size, coordinates, limits
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2031 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/battle2/battle2.h
@@ -58,7 +58,7 @@ namespace Battle2
 	u32  exp2;
     };
 
-    Result Loader(Army::army_t &, Army::army_t &, u16);
+    Result Loader(Army::army_t &, Army::army_t &, s32);
     void UpdateMonsterInfoAnimation(const std::string &);
     void UpdateMonsterAttributes(const std::string &);
