+++ fheroes2/src/fheroes2/kingdom/kingdom.h
@@ -110,11 +110,11 @@ public:
     void ActionNewWeek(void);
     void ActionNewMonth(void);
 
-    void SetVisited(const u16 index, const MP2::object_t object = MP2::OBJ_ZERO);
+    void SetVisited(const s32 index, const MP2::object_t object = MP2::OBJ_ZERO);
     u16  CountVisitedObjects(const MP2::object_t) const;
     bool isVisited(const u8 object) const;
     bool isVisited(const Maps::Tiles &) const;
-    bool isVisited(u16, u8) const;
+    bool isVisited(s32, u8) const;
 
     bool HeroesMayStillMove(void) const;
 
@@ -159,7 +159,7 @@ private:
     Recruits recruits;
 
     Castle *ai_capital;
-    std::map<u16, MP2::object_t> ai_objects;
+    std::map<s32, MP2::object_t> ai_objects;
 
     std::list<IndexObject> visit_object;
