+++ fheroes2/src/fheroes2/maps/ground.h
@@ -48,8 +48,8 @@ namespace Maps
         };
 
         const char* String(u16);
-        u16 GetBasePenalty(const u16 index, const u8 pathfinding);
-        u16 GetPenalty(const u16 index, const Direction::vector_t direct, const u8 pathfinding);
+        u16 GetBasePenalty(const s32 index, const u8 pathfinding);
+        u16 GetPenalty(const s32 index, const Direction::vector_t direct, const u8 pathfinding);
     }
 }
